requirements.txt
Flask == 0.10.1
Flask_Admin == 1.3.0
Flask-Security == 1.7.4
flask-mongoengine == 0.7.1
pymongo==2.7.2
Flask-MongoRest == 0.1.1
hashids == 1.1.0
flask-restful == 0.3.4
simple-crypt == 4.1.7
schema == 0.4.0
Flask-Login == 0.2.11
Flask-SSLify == 0.1.4
Flask-WTF == 0.9.5
oauth2 == 1.9.0.post1
http://github.com/pythonforfacebook/facebook-sdk/tarball/master
https://github.com/PrashntS/flask-oauth/archive/0.14.zip
Flask-OAuthlib == 0.9.2
https://github.com/PrashntS/flask-social/archive/1.6.3.zip
Click == 3.3
meinheld == 0.5.8
flask-script == 2.0.5
jsonschema == 2.5.1
https://github.com/PrashntS/Flask-Uploads-Py3/archive/0.1.4.zip
requests == 2.6.2